% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_filter_shortcuts}
\alias{get_filter_shortcuts}
\title{Get all available filter shortcuts}
\usage{
get_filter_shortcuts(lang = "ES", validate = TRUE, verbose = FALSE)
}
\arguments{
\item{lang}{(string): language. Set to 'ES' for the Spanish version of the
shortcuts or set to 'EN' for the English version of the shortcuts.}

\item{validate}{(logical): validate input parameters.}

\item{verbose}{(logical): print additional information.}
}
\value{
Data frame with information of the available filter shortcuts
}
\description{
Get all available filter shortcuts
}
\examples{
\dontrun{
get_filter_shortcuts()
get_filter_shortcuts(lang = "EN")
}

}
