% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata_operations}
\alias{get_metadata_operations}
\title{Get available operations}
\usage{
get_metadata_operations(
  operation = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{operation}{(string): code of the operation}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English.}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}
}
\value{
Data frame with information of the available operations
}
\description{
Get available operations
}
\examples{
\dontrun{
get_metadata_operations()
get_metadata_operations("IPC")
}

}
