% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata_periodicity}
\alias{get_metadata_periodicity}
\title{Get all possible periodicities or periodicities used in a operation}
\usage{
get_metadata_periodicity(
  operation = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{operation}{(string): Code of the operation}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English.}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}
}
\value{
Data frame with information of the available periodicities
}
\description{
Get all possible periodicities or periodicities used in a operation
}
\examples{
\dontrun{
get_metadata_periodicity()
get_metadata_periodicity(operation = "IPC")
}

}
