% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata_publication_dates}
\alias{get_metadata_publication_dates}
\title{Get the dates of a publication}
\usage{
get_metadata_publication_dates(
  publication = NULL,
  det = 0,
  tip = NULL,
  lang = "ES",
  page = 1,
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{publication}{(int): id of the publication}

\item{det}{(int): level of detail (0, 1 ,2)}

\item{tip}{(string): set to 'A' for friendly output, set to 'M' to include metadata or set to 'AM' for both}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English}

\item{page}{(int): page number. The retrieved result of the query is paginated (page=0 retrieves all pages)}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}
}
\value{
Data frame with information of the dates of a publication
}
\description{
Get the dates of a publication
}
\examples{
\dontrun{
get_metadata_publication_dates(publication = 8)
}

}
