% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata_publications}
\alias{get_metadata_publications}
\title{Get all the publications or the publications from an operation}
\usage{
get_metadata_publications(
  operation = NULL,
  det = 0,
  lang = "ES",
  page = 1,
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{operation}{(string): code of the operation}

\item{det}{(int): level of detail (0, 1 ,2)}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English}

\item{page}{(int): page number. The retrieved result of the query is paginated (page=0 retrieves all pages)}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}
}
\value{
Data frame with information of the publications
}
\description{
Get all the publications or the publications from an operation
}
\examples{
\dontrun{
get_metadata_publications()
get_metadata_publications(operation = "IPC")
}

}
