% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_series.R
\name{get_metadata_series_operation}
\alias{get_metadata_series_operation}
\title{Get all the series present in a operation}
\usage{
get_metadata_series_operation(
  operation = NULL,
  n = 500,
  det = 0,
  tip = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{operation}{(string): code of the operation}

\item{n}{(int): number of series to retrieve}

\item{det}{(int): level of detail (0, 1 ,2)}

\item{tip}{(string): set to 'A' for friendly output, set to 'M' to include metadata or set to 'AM' for both}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English.}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}
}
\value{
Data frame with information of the series belonging to an operation
}
\description{
Get all the series present in a operation
}
\examples{
\dontrun{
get_metadata_series_operation(operation = "IPC", n = 500)
}

}
