% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_series.R
\name{get_metadata_series_varval}
\alias{get_metadata_series_varval}
\title{Get metadata information of series belonging to an operation}
\usage{
get_metadata_series_varval(
  operation = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{operation}{(string): code of the operation}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}
}
\value{
Data frame with information about the variables and values that
define the series belonging to an operation
}
\description{
Get metadata information of series belonging to an operation
}
\examples{
\dontrun{
get_metadata_series_varval(operation = "IPC")
}

}
