% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tables.R
\name{get_metadata_tables_operation}
\alias{get_metadata_tables_operation}
\title{Get all the tables present in a operation}
\usage{
get_metadata_tables_operation(
  operation = NULL,
  det = 0,
  tip = NULL,
  geo = NULL,
  lang = "ES",
  page = 1,
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{operation}{(string): code of the operation}

\item{det}{(int): level of detail (0, 1 ,2)}

\item{tip}{(string): set to 'A' for friendly output, set to 'M' to include metadata or set to 'AM' for both}

\item{geo}{(int): set to 0 for national tables or set to 1 for tables with
a greater level of disaggregation}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English}

\item{page}{(int): page number. The retrieved result of the query is paginated}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}
}
\value{
Data frame with information of the available tables of an operation
}
\description{
Get all the tables present in a operation
}
\examples{
\dontrun{
get_metadata_tables_operation(operation = "IPC")
}

}
