% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata_values}
\alias{get_metadata_values}
\title{Get all possible values from a variable}
\usage{
get_metadata_values(
  operation = NULL,
  variable = NULL,
  det = 0,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{operation}{(string): code of the operation}

\item{variable}{(int): id of a variable}

\item{det}{(int): level of detail (0, 1 ,2)}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English.}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}
}
\value{
Data frame with information of the available values of a variable
}
\description{
Get all possible values from a variable
}
\examples{
\dontrun{
get_metadata_values(variable = 115)
get_metadata_values(operation = "IPC", variable = 115)
}

}
