% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata_variables}
\alias{get_metadata_variables}
\title{Get all possible variables or variables used in a operation}
\usage{
get_metadata_variables(
  operation = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{operation}{(string): Code of the operation}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English.}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}
}
\value{
Dataframe with information of the available variables
}
\description{
Get all possible variables or variables used in a operation
}
\examples{
get_metadata_variables()
get_metadata_variables(operation = "IPC")

}
