% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_series.R
\name{get_data_series}
\alias{get_data_series}
\title{Get data from series}
\usage{
get_data_series(
  codSeries = NULL,
  nlast = 1,
  dateStart = NULL,
  dateEnd = NULL,
  det = 0,
  tip = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE,
  unnest = FALSE
)
}
\arguments{
\item{codSeries}{(string): Code of the series}

\item{nlast}{(int): number of data or periods to retrieve}

\item{dateStart}{(string): Initial date of the requested data}

\item{dateEnd}{(string): Finish date of the requested data}

\item{det}{(int): Level of detail (0, 1 ,2)}

\item{tip}{(string): Set to 'A' for friendly output, set to 'M' to include metadata or set to 'AM' for both}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English.}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}

\item{unnest}{(logical): set to TRUE to obtain a single dataframe of data}
}
\value{
Dataframe with data of a series
}
\description{
Retrieve data from series published by INE calling the API
}
\examples{
\dontrun{
get_data_series(codSeries = "IPC251856")
get_data_series(codSeries = "IPC251856", nlast = 5)
get_data_series(codSeries = "IPC251856", dateStart = "2023/01/01", dateEnd = "2023/05/01")
}

}
