% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_series.R
\name{get_data_series_filter}
\alias{get_data_series_filter}
\title{Get data from series that fulfill the conditions of a filter}
\usage{
get_data_series_filter(
  operation = NULL,
  filter = NULL,
  period = NULL,
  nlast = 1,
  det = 0,
  tip = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE,
  unnest = FALSE,
  geocode = FALSE,
  shortcut = FALSE
)
}
\arguments{
\item{operation}{(string): Code of the operation}

\item{filter}{(list): list of variables and values, list(idvariable1 = idvalue1, idvariable2 = idvalue2)}

\item{period}{(int): id of the periodicity of the series (1:monthly, 3:quarterly, 12:annual)}

\item{nlast}{(int): number of data or periods to retreive}

\item{det}{(int): level of detail (0, 1 ,2)}

\item{tip}{(string): set to 'A' for friendly output, set to 'M' to include metadata or set to 'AM' for both}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English.}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}

\item{unnest}{(logical): obtain a single dataframe of data}

\item{geocode}{(logical): set to TRUE to obtain the code of national, ccaa, provinces or municipalities}

\item{shortcut}{(logical): enable the use of shortcut names in the filter}
}
\value{
Dataframe with data of series
}
\description{
Get data from series that fulfill the conditions of a filter
}
\examples{
get_data_series_filter(operation = "IPC", filter = list("115"= "29", "3" = "84", "762" = ""),
 period = 1)

}
