% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tables.R
\name{get_data_table}
\alias{get_data_table}
\title{Get data from a table}
\usage{
get_data_table(
  idTable = NULL,
  filter = NULL,
  nlast = NULL,
  det = NULL,
  tip = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE,
  unnest = FALSE,
  inecode = FALSE,
  extractmetadata = FALSE
)
}
\arguments{
\item{idTable}{(int): id of the table}

\item{filter}{(list): list of variables and values, list(idvariable1 = idvalue1, idvariable2 = idvalue2)}

\item{nlast}{(int): number of periods to retrieve}

\item{det}{(int): level of detail (0, 1 ,2)}

\item{tip}{(string): set to 'A' for friendly output, set to 'M' to include metadata or set to 'AM' for both}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English.}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}

\item{unnest}{(logical): set to TRUE to obtain a single data frame of data}

\item{inecode}{(logical): set to TRUE to obtain the code of national, ccaa, provinces or municipalities}

\item{extractmetadata}{(logical): set to TRUE to extract metadata to individual data frame columns}
}
\value{
Data frame with data of a table
}
\description{
Get data from a table
}
\examples{
\dontrun{
get_data_table(idTable = 50902)
get_data_table(idTable = 8105, filter = list("18"="454"), verbose = TRUE)
get_data_table(idTable = 33387,
filter = list(tipodematerial = c("extraccionnacional", "2mineralesmetalicosmineralenbruto")))
get_data_table(idTable = "t20/e245/p08/l0/01001.px",
filter = list(edad3gruposdeedad = "015anos", sexo = c("mujeres", "hombres")))
}

}
