% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_series.R
\name{get_metadata_series}
\alias{get_metadata_series}
\title{Get a series information}
\usage{
get_metadata_series(
  codSeries = NULL,
  det = 0,
  tip = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{codSeries}{(string): Code of the series}

\item{det}{(int): Level of detail (0, 1 ,2)}

\item{tip}{(string): Set to 'A' for friendly output, set to 'M' to include metadata or set to 'AM' for both}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English.}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}
}
\value{
Dataframe with information of a series
}
\description{
Get a series information
}
\examples{
get_metadata_series(codSeries = "IPC206449")

}
