% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_series.R
\name{get_metadata_series_filter}
\alias{get_metadata_series_filter}
\title{Get all the series that fulfill the conditions of a filter}
\usage{
get_metadata_series_filter(
  operation = NULL,
  filter = NULL,
  periodicity = NULL,
  det = 0,
  tip = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE,
  shortcut = FALSE
)
}
\arguments{
\item{operation}{(string): code of the operation}

\item{filter}{(list): list of variables and values, list(idvariable1 = idvalue1, idvariable2 = idvalue2)}

\item{periodicity}{(int): id of the periodicity of the series. Common periodicities:
1 (monthly), 3 (quarterly), 6 (bi-annual), 12 (annual).}

\item{det}{(int): level of detail (0, 1 ,2)}

\item{tip}{(string): set to 'A' for friendly output, set to 'M' to include metadata or set to 'AM' for both}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English.}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}

\item{shortcut}{(logical): enable the use of shortcut names in the filter}
}
\value{
Data frame with information of the series that fulfill the conditions of a filter
}
\description{
Get all the series that fulfill the conditions of a filter
}
\examples{
\dontrun{
get_metadata_series_filter(operation = "IPC", filter = list("115"= "29", "3" = "84", "762" = ""),
 periodicity = 1)
 }

}
