% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_series.R
\name{get_metadata_series_table}
\alias{get_metadata_series_table}
\title{Get all the series belonging to a table}
\usage{
get_metadata_series_table(
  idTable = NULL,
  filter = NULL,
  det = 0,
  tip = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE,
  metanames = FALSE,
  metacodes = FALSE
)
}
\arguments{
\item{idTable}{(int): code of the table}

\item{filter}{(list): list of variables and values, list(idvariable1 = idvalue1, idvariable2 = idvalue2)}

\item{det}{(int): level of detail (0, 1 ,2)}

\item{tip}{(string): set to 'A' for friendly output, set to 'M' to include metadata or set to 'AM' for both}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}

\item{metanames}{(logical): set to TRUE to extract the name of the values to individual data frame columns
from metadata information}

\item{metacodes}{(logical): set to TRUE to extract the codes and ids of the values to individual data frame columns
from metadata information}
}
\value{
Data frame with information of the series belonging to a table
}
\description{
Get all the series belonging to a table
}
\examples{
\dontrun{
get_metadata_series_table(idTable = 50902)
}

}
