% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_series.R
\name{get_metadata_series_values}
\alias{get_metadata_series_values}
\title{Get all the values used in a series}
\usage{
get_metadata_series_values(
  codSeries = NULL,
  det = 0,
  tip = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{codSeries}{(string): code of the series}

\item{det}{(int): level of detail (0, 1 ,2)}

\item{tip}{(string): set to 'A' for friendly output, set to 'M' to include metadata or set to 'AM' for both}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English.}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}
}
\value{
Data frame with information of the values of a series
}
\description{
Get all the values used in a series
}
\examples{
\dontrun{
get_metadata_series_values(codSeries = "IPC206449")
}

}
