% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tables.R
\name{get_metadata_table_groups}
\alias{get_metadata_table_groups}
\title{Get all groups for a specific a table}
\usage{
get_metadata_table_groups(
  idTable = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{idTable}{(int): id of the table. For further information about
ids click this \href{https://uvima.github.io/ineapir/articles/identify_codes.html}{link}.}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English.}

\item{validate}{(logical): validate input parameters.}

\item{verbose}{(logical): print additional information, including the URL to call the API service.}
}
\value{
Data frame with information of the groups according to the table specified in the function
}
\description{
Get all groups for a specific a table
}
\examples{
\dontrun{
get_metadata_table_groups(idTable = 50902)
}

}
