% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tables.R
\name{get_metadata_table_Values}
\alias{get_metadata_table_Values}
\title{Get all the values of a group of a table}
\usage{
get_metadata_table_Values(
  idTable = NULL,
  idGroup = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{idTable}{(int): code of the table}

\item{idGroup}{(int): code of the group of variables}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}
}
\value{
Data frame with information of the values of a group of a table
}
\description{
Get all the values of a group of a table
}
\examples{
\dontrun{
get_metadata_table_Values(idTable = 50902, idGroup = 110889)
}

}
