% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_tables.R
\name{get_metadata_table_varval}
\alias{get_metadata_table_varval}
\title{Get metadata information of a table}
\usage{
get_metadata_table_varval(
  idTable = NULL,
  lang = "ES",
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{idTable}{(int): code of the table}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English}

\item{validate}{(logical): validate the input parameters. A TRUE value implies less API calls}

\item{verbose}{(logical): print additional information}
}
\value{
Data frame with information about the variables and values that
define the table
}
\description{
Get metadata information of a table
}
\examples{
\dontrun{
get_metadata_table_varval(idTable = 50902)
}

}
