% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata_values}
\alias{get_metadata_values}
\title{Get all values for a specific variable}
\usage{
get_metadata_values(
  operation = NULL,
  variable = NULL,
  det = 0,
  lang = "ES",
  page = 0,
  validate = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{operation}{(string): code of the operation. Provide code to get all
the values for the given operation. To obtain a list of
available operations see \code{\link[=get_metadata_operations]{get_metadata_operations()}}.}

\item{variable}{(int): id of a variable. To obtain a list of
available variables see \code{\link[=get_metadata_variables]{get_metadata_variables()}}.}

\item{det}{(int): level of detail. Valid values: 0, 1 or 2.}

\item{lang}{(string): language of the retrieved data. Set to 'ES' for Spanish or set to 'EN' for English.}

\item{page}{(int): page number. The retrieved result of the query is paginated (page=0 retrieves all pages).}

\item{validate}{(logical): validate input parameters. A FALSE value means fewer API calls.}

\item{verbose}{(logical): print additional information, including the URL to call the API service.}
}
\value{
Data frame with information of the available values for the variable specified in the function
}
\description{
Get all values for a specific variable
}
\examples{
\dontrun{
get_metadata_values(variable = 115)
get_metadata_values(operation = "IPC", variable = 115)
}

}
